chore: resolve code review comments
